refactor: clean options in program
