+++ MimicWebApi/Controllers/TestController.cs
@@ -16,4 +16,11 @@ public class TestController(IUserService userService) : ControllerBase
         userService.Add(user.ToUser());
         return Ok();
     }
+
+    [HttpGet]
+    public IActionResult TestAuthWithUnBordered([FromBody] UserDTO user)
+    {
+        
+        return Ok();
+    }
 }
